+++ Runtime/Scripts/Actions/ActionWrapper.cs
@@ -1,0 +1,38 @@
+using System;
+using UnityEngine.InputSystem;
+
+namespace NPTP.InputSystemWrapper.Actions
+{
+    public abstract class ActionWrapper
+    {
+        internal InputAction InputAction { get; }
+        
+        private event Action<InputAction.CallbackContext> onEvent;
+        public event Action<InputAction.CallbackContext> OnEvent
+        {
+            add { onEvent -= value; onEvent += value; }
+            remove => onEvent -= value;
+        }
+
+        internal ActionWrapper(InputAction inputAction)
+        {
+            InputAction = inputAction;
+        }
+        
+        internal void RegisterCallbacks()
+        {
+            InputAction.started += HandleActionEvent;
+            InputAction.performed += HandleActionEvent;
+            InputAction.canceled += HandleActionEvent;
+        }
+        
+        internal void UnregisterCallbacks()
+        {
+            InputAction.started -= HandleActionEvent;
+            InputAction.performed -= HandleActionEvent;
+            InputAction.canceled -= HandleActionEvent;
+        }
+
+        private void HandleActionEvent(InputAction.CallbackContext context) => onEvent?.Invoke(context);
+    }
+}