+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Input System Wrapper
 ## Changelog
 
+1.1.0
+- Actions are now accessed via one more layer of wrapping. This makes more sense to read. E.g.: Input.UI.OnClick event becomes Input.UI.Click.OnEvent, to fit with other functionality inside Input.UI.Click.
+- ActionWrapper now supports buttons, values and pass-throughs, with explicitly-typed ReadValue methods.
+
 1.0.0
+- Initial version